clojure: Fix face name of fancified symbols
Fix issue #9271: Setting clojure-enable-fancify-symbols to t causes
the following to the be printed repeatedly in the message buffer:

    invalid face reference: t

The problem is that the clojure/fancify-symbols function adds font-lock
keywords with highlight forms containing face-name expressions that
evaluate to the return value of compose-region, which apparently is t;
this value is interpreted as a face name.

The solution to this problem is that the face-name expressions should
evaluate to nil, which will be interpreted as an empty list of properties
instead of a face name.

* layers/+lang/clojure/funcs.el (clojure/fancify-symbols): Define keywords
using face-name expressions that evaluate to nil.