Fix flyspell correction on spell-checking layer
Using the `spell-checking` layer mini-mode `Space S .` would cause an error

```
Wrong type argument: commandp -- flyspell-correct-previous-word-generic
```

According to the [README](https://github.com/d12frosted/flyspell-correct#deprecations-in-v05), it has since moved the command to a new name, which is not available anymore.

This commit changes the command used to point to the new name to make the transient mode work again.