FAQ: Set sentence-end-double-space in user-config
The FAQ has an entry for configuring Spacemacs to use double-space as the
sentence delimiter.  This FAQ entry instructs the reader to set the
sentence-end-double-space variable in the dotspacemacs/user-init function.

The problem is that the spacemacs-defaults layer sets
sentence-end-double-space when it loads, and Spacemacs loads layers after
calling dotspacemacs/user-init, which means the layer overrides any setting
for sentence-end-double-space in dotspacemacs/user-init.

To solve this problem, this commit modifies the FAQ entry to instruct the
reader to set the variable in the dotspacemacs/user-config function, which
Spacemacs runs after loading layers.

The FAQ entry was added in commit ee4ad69847f70bfc5159a744a4ddeaa42c7468c2.

* doc/FAQ.org: Instruct the reader to set sentence-end-double-space in
dotspacemacs/user-init rather than in dotspacemacs/user-config.