[org] Add agenda transient keybinding for org-agenda-show-clocking-issues
As documented upstream [1], the `org-agenda-show-clocking-issues` view can be
used to show overlapping clocks, gaps, etc. on the agenda, and is bound by
default to `v c`. This commit adds a keybinding it in the org-agenda transient mode.

This may not be the best place for it, because it isn't a proper toggle;
upstream says to clear the issues flags by toggling logbook mode itself and does
not provide a clocking issues toggle directly.

[1] https://orgmode.org/manual/Agenda-commands.html#Agenda-commands