[chinese][doc] Add space before open parenthesis
A space was added before the open parenthesis containing chinese characters, in
the section titles.

This solves a convention issue in the table of contents links (that are
generated from the titles), where the latin and chinese characters were next to
each other without a separating space.

When the space is added, then the generated links change the space to a dash.

A space was also added before any open parenthesis (with chinese characters) in
the non title text.

And a single full-width close parenthesis was changed to the half-width version
to match the other close parenthesis.

Thanks AmaiKinono for pointing out the issue.