dash: provide config option to skip auto-load of all docsets
In the upstream dash-docs and helm-dash documentation, they suggest a workflow
where users leave relatively little in `dash-common-docsets`, and instead add
per-mode hooks to enable specific docsets, since a user is unlikely to want
Python documentation while in a C++ buffer and vice versa. Currently, the
spacemacs dash layer will automatically discover all installed docsets and add
them to `dash-common-docsets`.

This commit adds a config option `dash-autoload-common-docsets` that can be set
to nil to skip the auto-loading and allow users to opt-in.

Co-Authored-By: duianto <otnaiud@gmail.com>