[helm] Fix rg interaction with its config file
If a user uses ripgrep persistent configuration file that sets options
helm overrides the argument parser for ripgrep might incorrectly parse
the command line used by helm.

The problem is encountered whenever an argument option occurs multiple
times and the last one is of this form: '--arg value' or '-a v'. If that
last argument is followed by the positional argument (input) then its
value will be treated as the input resulting in error due to too many
positional arguments.

In other worlds, this would fail to be parsed correctly:
    rg --max-columns=10 --max-columns 20 input
    rg --max-columns 10 --max-columns 20 input

While this works as expected:
    rg --max-columns=10 --max-columns=20 input
    rg --max-columns 10 --max-columns=20 input

The solution would be to use arguments in the form of '--arg=value'
which will never be treated as a positional argument.

Fixes #10953