Fix helm-spacemacs-help-faq, SPC h f
The function: helm-org-get-candidates was removed upstream.

Improvements:
Recenter the target candidate.

Reversed the helm headings list, to match the order of the
questions in FAQ.org, so that the Common questions are
listed first.

Removed the : (colon) after the heading: How do I
in FAQ.org, to match the other parent headings.